Separated components for manager table
