+++ src/components/manager/checkin/check-in-header.component.tsx
@@ -1,0 +1,28 @@
+import * as React from 'react';
+
+/*
+  *The check-in header component
+*/
+
+
+export class CheckInHeaderComponent extends React.Component<{}> {
+
+  public render() {
+    return (
+      <>
+      <thead>
+            <tr>
+              <th scope="col">User ID</th>
+              <th scope="col">First Name</th>
+              <th scope="col">Last Name</th>
+              <th scope="col">Cohort</th>
+              <th scope="col">Check-in Time</th>
+            </tr>
+          </thead>
+      </>
+
+    );
+  }
+}
+
+export default CheckInHeaderComponent
